requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
Flask>=2.0.0
Werkzeug>=2.0.0
